fix env_list
